Mimic setting `softtabstop` to -1 more accurately
When `shiftwidth` is 0, setting `softtabstop` to -1 does not disable it,
but the fallback code would have done just that by blindly assigning it
the value of `shiftwidth` -- i.e., zero.  In that case, assign the value
of `tabstop` instead.

Also, explicitly refer to local options where appropriate.  This doesn't
change any behavior but makes the intention clear.