Use U+2192 (→) in UTF-8 `showbreak`
According to the Vim help [*], "[only] printable single-cell characters
are allowed [in 'showbreak'], excluding <Tab> and comma".  U+21AA (↪)
appears to be multi-cell (at least on Apple Terminal 2.9.5 [421.2] with
10-pt. Adobe Source Code Pro), but I'm getting away with it by suffixing
a space, which Apple Terminal happens to not display opaquely.

Make "showbreak" more reliable by replacing U+21AA with the (apparently)
single-cell U+2192.  I don't think it looks as good, but I'll get used
to it.

  [*] As of Vim 9.0 with patches 1 through 1276.