Simplify setting options based on `encoding`
Considering how narrow its task is, the code that sets options based on
`encoding` is inelegant and hard to understand, and frankly it makes me
a little upset.  There is an API of sorts between vimrc.local and vimrc
-- the former assembles a data structure that the latter operates on --
but I designed it poorly, so the two files remain tightly coupled.  The
option values are expressed using double-quoted `\x..` escape sequences
to represent non-ASCII characters; this keeps the files fully ASCII and
thus `encoding`-agnostic [*], but it also makes them hard to understand
without meticulous comments mapping series of escape sequences to their
corresponding characters.  Finally: the code has reacted to mid-session
changes to `encoding` since it was first implemented in eaa0694 (vimrc:
Use Unicode listchars and add nbsp, 2016-02-22); this behavior directly
fosters much of the complexity I'm tired of, yet in practice `encoding`
only ever changes when I'm testing this code.

Replace everything with a dramatically simpler implementation that only
updates options once, at startup.  Set the options from a separate file
that's sourced if and only if `encoding` is "utf-8", permitting literal
UTF-8 characters to be used safely.  Replace `let` and `\x..` sequences
with `set` and UTF-8 characters; the latter is more legible and doesn't
require `+eval`.  (Extending this to values other than "utf-8" would be
trivial.  Maybe I'll actually do it one day.)

I initially wanted to improve the `encoding`-tracking behavior to avoid
updating options that the user had changed.  I reimplemented everything
in a separate autoload file, leaving a conditionalized function call in
vimrc and a flag variable in vimrc.local.  The new API was much better,
but the improved behavior needed a more complex data structure and more
code to manage it, and in the end I decided that the whole endeavor was
pointless because `encoding` never actually changes.

(And yes, I deliberately made every line in this commit message exactly
71 characters long, with the exception of the subject, the last line of
each paragraph, and the footnote.  The first lines I wrote were aligned
on the right by happenstance, and I challenged myself to keep it going.
What a tremendous waste of time.  I'd better not do it again.)

    [*] I don't know how `+ebcdic` affects this.  Who cares.