Set `encoding`-gated options for GUI independently
Since I implemented it in eaa0694 (vimrc: Use Unicode listchars and add
nbsp, 2016-02-22), the code that sets options according to the value of
`encoding` has never distinguished between GUI and non-GUI modes when
deciding whether to activate or not.  (Of course, until 3f8a22b [Stop
universally setting options based on encoding, 2021-08-17] it was always
active.)  This precludes using non-ASCII values in the GUI but not in
the terminal, or vice versa, without manipulating locale variables or
the like.

Accommodate sourcing set_opts.utf-8.vim in gvimrc.local in the same way
as in vimrc.local, allowing the options to be set differently in GUI and
non-GUI mode.