README: Fix fallback installation command
The command

	git clone url dir &&
	cd dir &&
	git submodule update --init --recursive

is meant to be an alternative to

	git clone --recursive url dir

for versions of Git that predate the `--recursive` option to `clone`.
Turns out it's not much of an alternative because -- surprise! -- the
`--recursive` options to `clone` and `submodule update` were added in
the same release.  (No, I don't have access to any Git this old.)

Drop `--recursive` and add a new command that initializes second-level
submodules.  I don't think there are any submodules on the third level,
but extending this workaround should be straightforward, if necessary.