README: Broaden installation instructions
Link to each commit that implements a required Git feature, justifying
how the instructions are split up by Git release.

Add tips for extremely old Git releases, noting the missing features
that must be worked around.

Add a tip for the lack of Git, linking to GitHub documentation about
downloading an archive from the website.

Replace the verb "clone" in prose, since cloning is not the only way to
obtain the code.