Be more precise about clearing/resetting options
The `backspace` and `showbreak` options are being reset, when the actual
intent is to clear them.  For these particular options the result is the
same, but there's no guarantee that that will always be true, and it is
better for the code to say what it means.

Clear `backspace` and `showbreak` instead of resetting them.  Update the
README to describe `set modeline&` and `set modelines&` more precisely.