Add notes on protecting the GUI from vimrc.local
If there are commands in vimrc.local that should not affect the GUI but
cannot be gated behind !has('gui_running') (presumably because they are
workarounds for a lack of +eval), then gvimrc.local must override them.

Make note of this requirement in README.markdown and comments.  Provide
example commands in gvimrc.local.sample.