Adapt options to `encoding` by default
Despite my tone in 3f8a22b (Stop universally setting options based on
encoding, 2021-08-17), I usually do want to add non-ASCII characters to
listchars and showbreak.  It's inconvenient to have to explicitly opt in
to this behavior.

Move the option-modifying code back into vimrc and gvimrc and gate it
with a flag variable (something I'd tried before but discarded in favor
of 3391146 [Simplify setting options based on `encoding`, 2025-01-09]).
Modify the options if the variable is truthy or nonexistent, making this
the default behavior.  It can still be disabled by assigning a falsy
value in vimrc.local or gvimrc.local, thus retaining the most important
innovation of 3f8a22b.

Add examples to vimrc.local.sample and gvimrc.local.sample demonstrating
how to disable the default behavior and work around a lack of +eval.

This reverts 3f8a22b in substance, but not in spirit!